update authz structure
