+++ framework/packages/abstract-sdk/src/apis/authz.rs
@@ -2,14 +2,16 @@
 //! This module provides functionality to interact with the authz module of CosmosSDK Chains.
 //! It allows for granting authorizations to perform actions on behalf of an account to other accounts.
 
-use cosmos_sdk_proto::cosmos::{bank, base, staking};
-use cosmos_sdk_proto::{cosmos::authz, traits::Message, Any};
+use cosmos_sdk_proto::{cosmos::authz, traits::Message};
 use cosmwasm_std::{Addr, Binary, Coin, CosmosMsg, Timestamp};
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
 
 use crate::features::AccountIdentification;
 use crate::AbstractSdkResult;
+
+use super::stargate::authz::{
+    AuthZAuthorization, AuthorizationType, GenericAuthorization, Policy, SendAuthorization,
+    StakeAuthorization,
+};
 
 /// An interface to the CosmosSDK AuthZ module which allows for granting authorizations to perform actions on behalf of one account to other accounts.
 pub trait AuthZInterface: AccountIdentification {
@@ -84,11 +86,12 @@ impl AuthZ {
         }
     }
 
-    fn grant(
+    /// Generate cosmwasm message for the AuthZAuthorization type
+    pub fn grant_authorization<A: AuthZAuthorization>(
         &self,
         grantee: &Addr,
         expiration: Option<Timestamp>,
-        authorization: impl AuthZAuthorization,
+        authorization: A,
     ) -> CosmosMsg {
         let msg = authz::v1beta1::MsgGrant {
             granter: self.granter().to_string(),
@@ -118,7 +121,7 @@ impl AuthZ {
     ) -> CosmosMsg {
         let generic = GenericAuthorization::new(msg_type_url);
 
-        self.grant(grantee, expiration, generic)
+        self.grant_authorization(grantee, expiration, generic)
     }
 
     /// Grants send authorization to a **grantee**.
@@ -136,7 +139,7 @@ impl AuthZ {
     ) -> CosmosMsg {
         let send = SendAuthorization::new(spend_limit);
 
-        self.grant(grantee, expiration, send)
+        self.grant_authorization(grantee, expiration, send)
     }
 
     /// Grants stake authorization to a **grantee**.
@@ -158,200 +161,17 @@ impl AuthZ {
     ) -> CosmosMsg {
         let stake = StakeAuthorization::new(max_tokens, authorization_type, validators);
 
-        self.grant(grantee, expiration, stake)
-    }
-}
-
-fn convert_stamp(stamp: Timestamp) -> prost_types::Timestamp {
-    prost_types::Timestamp {
-        seconds: stamp.seconds() as i64,
-        nanos: stamp.nanos() as i32,
-    }
-}
-
-fn convert_coins(coins: Vec<Coin>) -> Vec<base::v1beta1::Coin> {
-    coins.into_iter().map(convert_coin).collect()
-}
-
-fn convert_coin(coin: Coin) -> base::v1beta1::Coin {
-    base::v1beta1::Coin {
-        denom: coin.denom,
-        amount: coin.amount.to_string(),
-    }
-}
-
-/// Represents a generic authorization grant.
-#[derive(Serialize, Deserialize, Clone, PartialEq, Eq, JsonSchema)]
-pub struct GenericAuthorization {
-    /// Allowed msg type_url
-    pub msg_type_url: String,
-}
-
-impl GenericAuthorization {
-    /// Create new generic authorization
-    pub fn new(msg_type_url: String) -> Self {
-        Self { msg_type_url }
-    }
-}
-
-/// Represents send authorization grant
-#[derive(Serialize, Deserialize, Clone, PartialEq, Eq, JsonSchema)]
-pub struct SendAuthorization {
-    /// Allowed spend limit
-    pub spend_limit: Vec<Coin>,
-}
-
-impl SendAuthorization {
-    /// create new send authorization
-    pub fn new(spend_limit: Vec<Coin>) -> Self {
-        Self { spend_limit }
-    }
-}
-
-/// (de)serializable representation of [AuthorizationType](staking::v1beta1::AuthorizationType)
-#[derive(Serialize, Deserialize, Clone, Copy, PartialEq, Eq, JsonSchema)]
-pub enum AuthorizationType {
-    /// see [staking::v1beta1::AuthorizationType::Unspecified]
-    Unspecified,
-    /// see [staking::v1beta1::AuthorizationType::Delegate]
-    Delegate,
-    /// see [staking::v1beta1::AuthorizationType::Undelegate]
-    Undelegate,
-    /// see [staking::v1beta1::AuthorizationType::Redelegate]
-    Redelegate,
-}
-
-impl From<AuthorizationType> for staking::v1beta1::AuthorizationType {
-    fn from(value: AuthorizationType) -> Self {
-        use staking::v1beta1::AuthorizationType as StAuthT;
-        match value {
-            AuthorizationType::Unspecified => StAuthT::Unspecified,
-            AuthorizationType::Delegate => StAuthT::Delegate,
-            AuthorizationType::Undelegate => StAuthT::Undelegate,
-            AuthorizationType::Redelegate => StAuthT::Redelegate,
-        }
-    }
-}
-
-/// (de)serializable representation of [Policy](staking::v1beta1::stake_authorization::Policy)
-#[derive(Serialize, Deserialize, Clone, PartialEq, Eq, JsonSchema)]
-pub enum Policy {
-    /// see [staking::v1beta1::stake_authorization::Policy::AllowList]
-    AllowList(Vec<Addr>),
-    /// see [staking::v1beta1::stake_authorization::Policy::DenyList]
-    DenyList(Vec<Addr>),
-}
-
-impl From<Policy> for staking::v1beta1::stake_authorization::Policy {
-    fn from(value: Policy) -> Self {
-        use staking::v1beta1::stake_authorization::Policy as StPolicy;
-        use staking::v1beta1::stake_authorization::Validators;
-        match value {
-            Policy::AllowList(allow_list) => StPolicy::AllowList(Validators {
-                address: allow_list.into_iter().map(Addr::into_string).collect(),
-            }),
-            Policy::DenyList(deny_list) => StPolicy::DenyList(Validators {
-                address: deny_list.into_iter().map(Addr::into_string).collect(),
-            }),
-        }
-    }
-}
-
-/// Represents stake authorization grant
-#[derive(Serialize, Deserialize, Clone, PartialEq, Eq, JsonSchema)]
-pub struct StakeAuthorization {
-    /// see [staking::v1beta1::StakeAuthorization::max_tokens]
-    pub max_tokens: Option<Coin>,
-    /// see [staking::v1beta1::StakeAuthorization::authorization_type]
-    pub authorization_type: AuthorizationType,
-    /// see [staking::v1beta1::StakeAuthorization::validators]
-    pub validators: Option<Policy>,
-}
-
-impl StakeAuthorization {
-    /// create new send authorization
-    pub fn new(
-        max_tokens: Option<Coin>,
-        authorization_type: AuthorizationType,
-        validators: Option<Policy>,
-    ) -> Self {
-        Self {
-            max_tokens,
-            authorization_type,
-            validators,
-        }
-    }
-}
-trait AuthZAuthorization {
-    type ProtoType: Message;
-
-    const TYPE_URL: &'static str;
-
-    /// Get `Any`
-    fn to_any(&self) -> Any {
-        Any {
-            type_url: Self::TYPE_URL.to_owned(),
-            value: self.to_proto().encode_to_vec(),
-        }
-    }
-
-    /// Get `Self::ProtoType`
-    fn to_proto(&self) -> Self::ProtoType;
-
-    fn grant(&self, expiration: Option<Timestamp>) -> authz::v1beta1::Grant {
-        authz::v1beta1::Grant {
-            authorization: Some(self.to_any()),
-            expiration: expiration.map(convert_stamp),
-        }
-    }
-}
-
-impl AuthZAuthorization for GenericAuthorization {
-    type ProtoType = authz::v1beta1::GenericAuthorization;
-
-    const TYPE_URL: &'static str = "/cosmos.authz.v1beta1.GenericAuthorization";
-
-    fn to_proto(&self) -> Self::ProtoType {
-        authz::v1beta1::GenericAuthorization {
-            msg: self.msg_type_url.clone(),
-        }
-    }
-}
-
-impl AuthZAuthorization for SendAuthorization {
-    type ProtoType = bank::v1beta1::SendAuthorization;
-
-    const TYPE_URL: &'static str = "/cosmos.bank.v1beta1.SendAuthorization";
-
-    fn to_proto(&self) -> Self::ProtoType {
-        bank::v1beta1::SendAuthorization {
-            spend_limit: convert_coins(self.spend_limit.clone()),
-        }
-    }
-}
-
-impl AuthZAuthorization for StakeAuthorization {
-    type ProtoType = staking::v1beta1::StakeAuthorization;
-
-    const TYPE_URL: &'static str = "/cosmos.staking.v1beta.StakeAuthorization";
-
-    fn to_proto(&self) -> Self::ProtoType {
-        let authorization_type: staking::v1beta1::AuthorizationType =
-            self.authorization_type.into();
-        staking::v1beta1::StakeAuthorization {
-            max_tokens: self.max_tokens.clone().map(convert_coin),
-            authorization_type: authorization_type.into(),
-            validators: self.validators.clone().map(Into::into),
-        }
+        self.grant_authorization(grantee, expiration, stake)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::mock_module::*;
+    use crate::{apis::stargate::convert_stamp, mock_module::*};
 
     use cosmwasm_std::testing::mock_dependencies;
+    use prost_types::Any;
 
     #[test]
     fn generic_authorization() {
