+++ framework/packages/abstract-sdk/src/apis/stargate/mod.rs
@@ -1,21 +1,22 @@
+pub mod authz;
 pub mod feegrant;
 
-use cosmos_sdk_proto::{
-    cosmos::base,
-    traits::{Message, Name},
-};
+use cosmos_sdk_proto::{cosmos::base, traits::Message};
 use cosmwasm_std::{Coin, Timestamp};
 use prost_types::Any;
 
 /// This trait allows generate `Any` and proto message from Stargate API message
 pub trait StargateMessage {
     /// Returned proto type
-    type ProtoType: Message + Name + Sized;
+    type ProtoType: Message;
+
+    // TODO: replacable by `Name` trait for ProtoType when authz gets it
+    fn type_url() -> String;
 
     /// Get `Any`
     fn to_any(&self) -> Any {
         Any {
-            type_url: Self::ProtoType::type_url(),
+            type_url: Self::type_url(),
             value: self.to_proto().encode_to_vec(),
         }
     }
